requirements.txt
#Original Modules
django >3.2,<3.3
python-decouple==3.5
Pillow==9.0.0
djangorestframework
coreapi

#Third Party Modules
markdown  
django-filter
drf-yasg[validation]
djangorestframework-simplejwt
# Email third party modules
django-mail-templated
#Deployment Modules
djoser

# refactoring and test
black
flake8